requirements.txt
fastapi>=0.115.0
strawberry-graphql[fastapi]>=0.262.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
